♻️ 時間表示の機能を統括 #47
